requirements.txt
pandas
numpy
scikit-learn
matplotlib
paho-mqtt
streamlit
